kitakami: camera: Enable sonycamera proprietary capabilities
Sony camera blobs are encrypted, so they need
com.sonyericsson.permission.CAMERA_EXTENDED permission from original
camera to be present in apk to enable Sony custom options (4k video, iso
settings, etc) camera.qcom.so checks if this permission is available in
current camera app but it can be bypassed when "ro.build.type=eng" thus
enabling Sony camera options. The point of this shim is to make
camera.qcom.so believe that we're currently on an engineering android
rom so it bypasses the permission check allowing camera apps to have
full capabilities.

Change-Id: Ic3c935cf8776a517781a5b6cca8f700afd083270